requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
faiss-cpu
numpy
unstructured[pdf]
unstructured-inference
python-dotenv
